# Improve free symbols search for large tensors