--- a/lambeq/backend/tensor.py
+++ b/lambeq/backend/tensor.py
@@ -242,7 +242,11 @@ class Box(grammar.Box):
             if isinstance(data, Mapping):
                 data = data.values()
             if isinstance(data, Iterable):
-                if not hasattr(data, 'shape') or data.shape != ():
+                if (
+                    not hasattr(data, 'shape') or data.shape != ()
+                ) and (
+                    not hasattr(data, 'dtype') or data.dtype != object
+                ):
                     return set().union(*map(recursive_free_symbols, data))
             # Remove scale before adding to set
             return {data.unscaled} if isinstance(data, Symbol) else set()