--- a/lambeq/training/nelder_mead_optimizer.py
+++ b/lambeq/training/nelder_mead_optimizer.py
@@ -259,7 +259,7 @@ class NelderMeadOptimizer(Optimizer):
                 raise ValueError(
                     'Objective function must return a scalar'
                 ) from e
-        return result
+        return result   # type: ignore[return-value]
 
     def backward(self, batch: tuple[Iterable[Any], np.ndarray]) -> float:
         """Calculate the gradients of the loss function.